Ensure new window that fails to become the foreground window is placed at the end of the interaction list
This is necessary in order to avoid messing with the order of the windows displayed
by the switcher. It is discussed in-depth in #1084 and worked upon in d350b0e and 8ab7a45.

Previously, we set the timestamp of any window that failed this test to the time the application
was started. That meant that newly spawned windows that failed the test had the same
timestamp until interacted with, so they were displayed according to the Z-order in
the switcher, so the previous logic only worked properly for the first window. This fix
addresses that.

The interaction list is obtained by sorting descending by `ft` the `htshwnds` list of the switcher
and then picking only the windows suitable for Alt-Tab. The switcher so computes and then
displays this list when invoked.