Try to maintain selection geography when the interaction list changes and the switcher is shown
Previously, when the interaction list changed while the switcher was displayed,
for example, by closing one of the displayed windows, or having some window
spawn on the desktop, the currently selected item in the switcher would reset,
using the same logic as if the switcher had just been triggered using `Alt`-`Tab`.

This patch addresses that shortcoming:
* If the previously selected window still exists, it will be selected in the newly
recomputed layout.
* If the previously selected window does not exist anymore (for example, the
user closed it from the switcher), then the program iterates over the old layout
from the index of the previously selected window and identifies the first window
there that still exists in the new layout and selects that.