Group windows primarly by `AppUserModelIDs`, when necessary
`AppUserModelIDs` (AUMIDs) are the official OS mechanism that groups together
entities like application windows. This commit implements the following
infrastructure:

* When displaying the interaction list for individual apps (`Alt` + `key above Tab`),
similar windows are grouped based on their AUMIDs being the same. If the
focused window has an AUMID, it will be grouped only with other windows that
share the same AUMID. If the focused window does not have an AUMID, then the
legacy behavior is employed (it looks for a match between the process IDs or
image paths', except for `rundll32` which is always ungrouped). This also implements
grouping for UWP apps, which were ungrouped previously.
* When the switcher is set to group windows of the same app, the behavior is now
similar to what is described above and it's coded in a single place within the source
code. Elsewhere, there's now a linked list that starts from the window that is
displayed in the switcher and reveals all the other windows that are grouped "under"
it, computed only once, when the layout is created. Similarly, the count of related
windows is computed only once and stored in the information of the window
displayed in the switcher.
* This also fixes a bug where having two windows of the same app open on two
different monitors with the settings to group windows of the same app and to
display windows onlt from the monitor containing the mouse enabled resulted in
the switcher displaying that app only on the monitor containing the window with
which the user interacted most recently with.
* Since the switcher now stores AUMIDs in the window information structures,
when it groups windows by app, it now queries the `shell:AppsFolder` and extracts
the app name from there, based on the AUMID. If the app does not have an AUMID
or a name cannot be obtained by this means, the legacy behavior is employed
(the file description of the application's executable is displayed).