Support switching between individual and global interaction lists when the switcher is shown
In line with other window switchers that implement such functionality, this
patch adds the following behavior:

* When displaying the global interaction list (the one which contains all
windows), pressing `Alt`+`key above Tab` will now switch to the individual
interaction list for the selected application. The same window which was
selected in the global interaction list will be selected in the individual list.
Previously, the behavior was to continue cycling the list.
* When displaying the individual interaction list (the one which contains
windows belonging to the same app), pressing `Alt`+`Tab` will now switch
to the global interaction list. The same window which was selected in the
individual list will be selected in the global list, except when grouping,
which will have the switcher select the corresponding group.
* This patch also fixes a bug that caused the switcher not to scroll the
window into view in 2 circumstances:
** The user closed a window but the next one doesn't fit in the available
space. The behavior is now corrected, the switcher brings the selection into
view.
** The user switches to the global list with a selection from the individual
list which doesn't fit in the initial view of the global list. This behavior is now
corrected.