Fixed a bug that still considered about to be closed windows when building layouts
This was a very nasty one, it messed up with the order of the windows.
For some windows, the "window removed" notification is received before
the actual window is removed from the desktop. On "window removal",
the switcher deletes the window from the timestamped window collection,
but the layout code still took into consideration non-timestamped but
otherwise valid Alt-Tab windows in order to spare one enumeration of
the Alt-Tab windows when the program started. Non-timestamped
windows were always placed at the end of the list. This created this terrible,
yet interesting bug, where right at startup, when few of the windows were
timestamped, about to be closed windows were spawning in the layout right
after the few timestamped windows and messing the decision of which window
to select (since the selected one is about to be closed). They popped up at the
fron of the non-timestamped windows because the window enumeration
API presents windows in the Z-order, and the about to be closed window
was just interacted with. Took quite a while to troubleshoot and it's
cumbersome to explain.

The solution was to enumerate Alt-Tab windows at startup, in order to have a
proper set of timestamped windows from the beginning. Furthermore, when
building the window list for the layout, the logic now discards valid windows
which are not timestamped, since they probably are about to be closed by
the system and we just got the notification, removed them from the
timestamped window list, but the system has yet to remove it yet and we were
requested a relayout, so we enumerate the windows and receieve this "ghost"
one as well.