Do not switch to windows displaying popups when attempting to close them
It seems the OS behavior is to keep the switcher open, and have the user
decide whether he/she actually wants to switch to the window that
requires attention. It's rather hard to miss since usually such windows either
flash their taskbar button, so they will flash in the switcher as well (example:
unsaved Notepad document, unsaved Wordpad document), either play a
sound, like the beep tone (exmaple: This PC -right click C: - Properties -
Tools - Error checking - Check).

I disabled the in-house custom behavior because it was inconsistent and racy:
it the timeout elapsed but the window did not process the close message,
the switcher would remain open, creating inconsistency, when in other
situations the app to be closed would display the popup before the timeout
in sws (300 ms) elapsed.