Made a custom global `GetLastActivePopup` function
This allows for easily changing the implementation if anything is required
besides what the standard method offered.

`GetLastActivePopup` doesn't work when the window we request for is
owned by another window. That seemed to be the case for at least a certain
case, and this seems to fix it. The switcher previously failed to switch to the
"Error Checking" pop-up in This PC - right click C: - Properties -
Tools - Error checking - Check.